Extract `builtin_helpers.rs`
