+++ compiler/noirc_frontend/src/hir/comptime/interpreter/builtin.rs
@@ -4,6 +4,11 @@ use std::{
 };
 
 use acvm::{AcirField, FieldElement};
+use builtin_helpers::{
+    check_argument_count, check_one_argument, check_three_arguments, check_two_arguments,
+    get_function_def, get_quoted, get_slice, get_trait_constraint, get_trait_def, get_type,
+    get_u32, hir_pattern_to_tokens,
+};
 use chumsky::Parser;
 use iter_extended::{try_vecmap, vecmap};
 use noirc_errors::Location;
@@ -12,15 +17,15 @@ use rustc_hash::FxHashMap as HashMap;
 use crate::{
     ast::IntegerBitSize,
     hir::comptime::{errors::IResult, value::add_token_spans, InterpreterError, Value},
-    hir_def::stmt::HirPattern,
     macros_api::{NodeInterner, Signedness},
-    node_interner::{FuncId, TraitId},
     parser,
     token::Token,
     QuotedType, Shared, Type,
 };
 
 use super::Interpreter;
+
+pub(crate) mod builtin_helpers;
 
 impl<'local, 'context> Interpreter<'local, 'context> {
     pub(super) fn call_builtin(
@@ -79,163 +84,8 @@ impl<'local, 'context> Interpreter<'local, 'context> {
     }
 }
 
-pub(super) fn check_argument_count(
-    expected: usize,
-    arguments: &[(Value, Location)],
-    location: Location,
-) -> IResult<()> {
-    if arguments.len() == expected {
-        Ok(())
-    } else {
-        let actual = arguments.len();
-        Err(InterpreterError::ArgumentCountMismatch { expected, actual, location })
-    }
-}
-
-pub(super) fn check_one_argument(
-    mut arguments: Vec<(Value, Location)>,
-    location: Location,
-) -> IResult<Value> {
-    check_argument_count(1, &arguments, location)?;
-
-    Ok(arguments.pop().unwrap().0)
-}
-
-pub(super) fn check_two_arguments(
-    mut arguments: Vec<(Value, Location)>,
-    location: Location,
-) -> IResult<(Value, Value)> {
-    check_argument_count(2, &arguments, location)?;
-
-    let argument2 = arguments.pop().unwrap().0;
-    let argument1 = arguments.pop().unwrap().0;
-
-    Ok((argument1, argument2))
-}
-
-pub(super) fn check_three_arguments(
-    mut arguments: Vec<(Value, Location)>,
-    location: Location,
-) -> IResult<(Value, Value, Value)> {
-    check_argument_count(3, &arguments, location)?;
-
-    let argument3 = arguments.pop().unwrap().0;
-    let argument2 = arguments.pop().unwrap().0;
-    let argument1 = arguments.pop().unwrap().0;
-
-    Ok((argument1, argument2, argument3))
-}
-
 fn failing_constraint<T>(message: impl Into<String>, location: Location) -> IResult<T> {
     Err(InterpreterError::FailingConstraint { message: Some(message.into()), location })
-}
-
-pub(super) fn get_array(
-    interner: &NodeInterner,
-    value: Value,
-    location: Location,
-) -> IResult<(im::Vector<Value>, Type)> {
-    match value {
-        Value::Array(values, typ) => Ok((values, typ)),
-        value => {
-            let type_var = Box::new(interner.next_type_variable());
-            let expected = Type::Array(type_var.clone(), type_var);
-            let actual = value.get_type().into_owned();
-            Err(InterpreterError::TypeMismatch { expected, actual, location })
-        }
-    }
-}
-
-fn get_slice(
-    interner: &NodeInterner,
-    value: Value,
-    location: Location,
-) -> IResult<(im::Vector<Value>, Type)> {
-    match value {
-        Value::Slice(values, typ) => Ok((values, typ)),
-        value => {
-            let type_var = Box::new(interner.next_type_variable());
-            let expected = Type::Slice(type_var);
-            let actual = value.get_type().into_owned();
-            Err(InterpreterError::TypeMismatch { expected, actual, location })
-        }
-    }
-}
-
-pub(super) fn get_field(value: Value, location: Location) -> IResult<FieldElement> {
-    match value {
-        Value::Field(value) => Ok(value),
-        value => {
-            let actual = value.get_type().into_owned();
-            Err(InterpreterError::TypeMismatch { expected: Type::FieldElement, actual, location })
-        }
-    }
-}
-
-pub(super) fn get_u32(value: Value, location: Location) -> IResult<u32> {
-    match value {
-        Value::U32(value) => Ok(value),
-        value => {
-            let expected = Type::Integer(Signedness::Unsigned, IntegerBitSize::ThirtyTwo);
-            let actual = value.get_type().into_owned();
-            Err(InterpreterError::TypeMismatch { expected, actual, location })
-        }
-    }
-}
-
-fn get_function_def(value: Value, location: Location) -> IResult<FuncId> {
-    match value {
-        Value::FunctionDefinition(id) => Ok(id),
-        value => {
-            let expected = Type::Quoted(QuotedType::FunctionDefinition);
-            let actual = value.get_type().into_owned();
-            Err(InterpreterError::TypeMismatch { expected, actual, location })
-        }
-    }
-}
-
-fn get_trait_constraint(value: Value, location: Location) -> IResult<(TraitId, Vec<Type>)> {
-    match value {
-        Value::TraitConstraint(trait_id, generics) => Ok((trait_id, generics)),
-        value => {
-            let expected = Type::Quoted(QuotedType::TraitConstraint);
-            let actual = value.get_type().into_owned();
-            Err(InterpreterError::TypeMismatch { expected, actual, location })
-        }
-    }
-}
-
-fn get_trait_def(value: Value, location: Location) -> IResult<TraitId> {
-    match value {
-        Value::TraitDefinition(id) => Ok(id),
-        value => {
-            let expected = Type::Quoted(QuotedType::TraitDefinition);
-            let actual = value.get_type().into_owned();
-            Err(InterpreterError::TypeMismatch { expected, actual, location })
-        }
-    }
-}
-
-fn get_type(value: Value, location: Location) -> IResult<Type> {
-    match value {
-        Value::Type(typ) => Ok(typ),
-        value => {
-            let expected = Type::Quoted(QuotedType::Type);
-            let actual = value.get_type().into_owned();
-            Err(InterpreterError::TypeMismatch { expected, actual, location })
-        }
-    }
-}
-
-fn get_quoted(value: Value, location: Location) -> IResult<Rc<Vec<Token>>> {
-    match value {
-        Value::Quoted(tokens) => Ok(tokens),
-        value => {
-            let expected = Type::Quoted(QuotedType::Quoted);
-            let actual = value.get_type().into_owned();
-            Err(InterpreterError::TypeMismatch { expected, actual, location })
-        }
-    }
 }
 
 fn array_len(
@@ -931,71 +781,4 @@ pub(crate) fn extract_option_generic_type(typ: Type) -> Type {
     assert_eq!(struct_type.name.0.contents, "Option");
 
     generics.pop().expect("Expected Option to have a T generic type")
-}
-
-fn hir_pattern_to_tokens(interner: &NodeInterner, hir_pattern: &HirPattern) -> Vec<Token> {
-    let mut tokens = Vec::new();
-    gather_hir_pattern_tokens(interner, hir_pattern, &mut tokens);
-    tokens
-}
-
-fn gather_hir_pattern_tokens(
-    interner: &NodeInterner,
-    hir_pattern: &HirPattern,
-    tokens: &mut Vec<Token>,
-) {
-    match hir_pattern {
-        HirPattern::Identifier(hir_ident) => {
-            let name = interner.definition_name(hir_ident.id).to_string();
-            tokens.push(Token::Ident(name));
-        }
-        HirPattern::Mutable(pattern, _) => {
-            tokens.push(Token::Keyword(crate::token::Keyword::Mut));
-            gather_hir_pattern_tokens(interner, pattern, tokens);
-        }
-        HirPattern::Tuple(patterns, _) => {
-            tokens.push(Token::LeftParen);
-            for (index, pattern) in patterns.iter().enumerate() {
-                if index != 0 {
-                    tokens.push(Token::Comma);
-                }
-                gather_hir_pattern_tokens(interner, pattern, tokens);
-            }
-            tokens.push(Token::RightParen);
-        }
-        HirPattern::Struct(typ, fields, _) => {
-            let Type::Struct(struct_type, _) = typ else {
-                panic!("Expected type to be a struct");
-            };
-
-            let name = struct_type.borrow().name.to_string();
-            tokens.push(Token::Ident(name));
-
-            tokens.push(Token::LeftBrace);
-            for (index, (field_name, pattern)) in fields.iter().enumerate() {
-                if index != 0 {
-                    tokens.push(Token::Comma);
-                }
-
-                let field_name = &field_name.0.contents;
-                tokens.push(Token::Ident(field_name.to_string()));
-
-                // If we have a pattern like `Foo { x }`, that's internally represented as `Foo { x: x }` so
-                // here we check if the field name is the same as the pattern and, if so, omit the `: x` part.
-                let field_name_is_same_as_pattern = if let HirPattern::Identifier(pattern) = pattern
-                {
-                    let pattern_name = interner.definition_name(pattern.id);
-                    field_name == pattern_name
-                } else {
-                    false
-                };
-
-                if !field_name_is_same_as_pattern {
-                    tokens.push(Token::Colon);
-                    gather_hir_pattern_tokens(interner, pattern, tokens);
-                }
-            }
-            tokens.push(Token::RightBrace);
-        }
-    }
 }
